requirements.txt
torch
torchvision
streamlit
pillow
urllib3